simple-definition-list-blocks/term の構成オブジェクトを修正
